+++ CHANGELOG.md
@@ -5,13 +5,13 @@
 
 ### Improvements
 
-- Display more Unicode information in the character section in the document inspector.
-- Improve trimming whitespaces in extracted outline items.
+- Provide more Unicode information about the selected character in the document inspector.
+- Improve the trimming of whitespaces in the extracted outline labels.
 
 
 ### Fixes
 
-- Fix an issue that the end point of the selection shifted after the hanging indent of a wrapped line when the line endings are CRLF and the selection is expanded by using the `⇧↑` or `⇧↓` shortcuts.
+- Fix an issue that the end point of a selection shifted after a hanging indent on a wrapped line when the line endings are CRLF and the selection is expanded using the `⇧↑` or `⇧↓` shortcuts.