+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixes
 
+- Fix an issue that the shortcut key for cycling through windows was not functioning correctly for certain input methods.
 - Fix an issue that the end point of a selection shifted after a hanging indent on a wrapped line when the line endings are CRLF and the selection is expanded using the `⇧↑` or `⇧↓` shortcuts.